:truck: chore: organize packages
