+++ lib/di/injection.dart
@@ -7,14 +7,14 @@ import '../data/network/source/local/movies/local_data_source_impl.dart';
 import '../data/network/source/local/movies/local_data_source_movies.dart';
 import '../data/network/source/local/series/local_data_source_impl.dart';
 import '../data/network/source/local/series/local_data_source_series.dart';
-import '../data/network/source/remote/movie_data_source.dart';
-import '../data/network/source/remote/movie_data_source_impl.dart';
-import '../data/network/source/remote/series_data_source.dart';
-import '../data/network/source/remote/series_data_source_impl.dart';
+import '../data/network/source/remote/movies/movie_data_source.dart';
+import '../data/network/source/remote/movies/movie_data_source_impl.dart';
+import '../data/network/source/remote/series/series_data_source.dart';
+import '../data/network/source/remote/series/series_data_source_impl.dart';
 import '../data/repository/movie_repository_impl.dart';
 import '../data/repository/series_repository_impl.dart';
-import '../data/service/movie_service_api.dart';
-import '../data/service/series_service_api.dart';
+import '../data/service/movies/movie_service_api.dart';
+import '../data/service/series/series_service_api.dart';
 import '../domain/repository/movie_repository.dart';
 import '../domain/repository/series_repository.dart';
 import '../domain/usecase/movies/add_to_favorites_use_case.dart';
