+++ lib/main.dart
@@ -3,10 +3,10 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:r6_moovie_app/presenter/bloc/favorites/favorite_bloc.dart';
 import 'package:r6_moovie_app/presenter/bloc/movies/movie_bloc.dart';
 import 'package:r6_moovie_app/presenter/bloc/series/series_bloc.dart';
-import 'package:r6_moovie_app/presenter/pages/favorites_screen.dart';
+import 'package:r6_moovie_app/presenter/pages/favorites/favorites_screen.dart';
 import 'package:r6_moovie_app/presenter/pages/main_screen.dart';
 import 'package:r6_moovie_app/presenter/pages/splash_screen.dart';
-import 'package:r6_moovie_app/presenter/widgets/home/app_bar_main.dart';
+import 'package:r6_moovie_app/presenter/widgets/comon/app_bar_main.dart';
 import 'package:r6_moovie_app/presenter/widgets/home/nav_bar_main.dart';
 import 'package:r6_moovie_app/presenter/widgets/home/search_bar_app.dart';
 import 'package:r6_moovie_app/resources/app_colors.dart';
