+++ lib/presenter/pages/favorites/favorites_series_screen.dart
