+++ lib/presenter/pages/movies/movies_details_screen.dart
@@ -1,36 +1,35 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:r6_moovie_app/presenter/pages/series/favorites_series_screen.dart';
-import 'package:r6_moovie_app/resources/app_colors.dart';
-import 'package:r6_moovie_app/utils/utils.dart';
 import 'package:r6_moovie_app/presenter/widgets/details/cast.dart';
 import 'package:r6_moovie_app/presenter/widgets/details/review.dart';
+import 'package:r6_moovie_app/resources/app_colors.dart';
+import 'package:r6_moovie_app/resources/app_strings.dart';
+import 'package:r6_moovie_app/utils/utils.dart';
 
-import '../../domain/entities/series.dart';
-import '../../resources/app_strings.dart';
-import '../bloc/favorites/favorite_bloc.dart';
-import '../bloc/favorites/favorite_state.dart';
-import '../widgets/details/info_row.dart';
-import '../widgets/details/media_detail_header.dart';
-import '../widgets/details/overview.dart';
-import '../widgets/home/favorite_toggle_button.dart';
+import '../../../domain/entities/movie.dart';
+import '../../bloc/favorites/favorite_bloc.dart';
+import '../../bloc/favorites/favorite_state.dart';
+import '../../widgets/details/info_row.dart';
+import '../../widgets/details/media_detail_header.dart';
+import '../../widgets/details/overview.dart';
+import '../../widgets/comon/favorite_toggle_button.dart';
+import '../favorites/favorites_screen.dart';
 
-class SeriesDetailsScreen extends StatefulWidget {
+class MovieDetailsScreen extends StatefulWidget {
   final dynamic item;
 
-  const SeriesDetailsScreen({super.key, required this.item});
+  const MovieDetailsScreen({super.key, required this.item});
 
   @override
-  // ignore: library_private_types_in_public_api
-  _SeriesDetailsScreenState createState() => _SeriesDetailsScreenState();
+  _MovieDetailsScreenState createState() => _MovieDetailsScreenState();
 }
 
-class _SeriesDetailsScreenState extends State<SeriesDetailsScreen> {
+class _MovieDetailsScreenState extends State<MovieDetailsScreen> {
   int _selectedIndex = 0;
 
   @override
   Widget build(BuildContext context) {
-    final Series series = widget.item;
+    final Movie movie = widget.item;
 
     return Scaffold(
       appBar: AppBar(
@@ -48,7 +47,7 @@ class _SeriesDetailsScreenState extends State<SeriesDetailsScreen> {
               Navigator.push(
                 context,
                 MaterialPageRoute(
-                    builder: (context) => FavoritesSeriesScreen()),
+                    builder: (context) => const FavoritesScreen()),
               );
             },
           ),
@@ -56,13 +55,10 @@ class _SeriesDetailsScreenState extends State<SeriesDetailsScreen> {
       ),
       body: BlocBuilder<FavoriteBloc, FavoriteState>(
         builder: (context, state) {
-          // ignore: unused_local_variable
           bool isFavorite = false;
           if (state is FavoritesLoadedState) {
             isFavorite =
-                state.favoriteSeries.any((series) => series.id == series.id);
-            isFavorite =
-                state.favoriteSeries.any((series) => series.id == series.id);
+                state.favoriteMovies.any((movie) => movie.id == movie.id);
           }
 
           return SingleChildScrollView(
@@ -72,17 +68,16 @@ class _SeriesDetailsScreenState extends State<SeriesDetailsScreen> {
                 Padding(
                   padding: const EdgeInsets.only(bottom: 20),
                   child: MediaDetailHeader(
-                    media: series,
+                    media: movie,
                     height: 250,
-                    action: FavoriteToggleButton(media: series),
+                    action: FavoriteToggleButton(media: movie),
                   ),
                 ),
                 InfoRow(
-                  releaseDate: Utils.formatDateToBrazilian(series.firstAirDate),
-                  vote: series.voteAverage.toString(),
-                  popularity: series.name,
-                ),
-                const SizedBox(height: 10),
+                    releaseDate: Utils.formatDateToBrazilian(movie.releaseDate),
+                    vote: movie.voteCount.toString(),
+                    popularity: movie.popularity.toString()),
+                const SizedBox(height: 5),
                 Row(
                   mainAxisAlignment: MainAxisAlignment.spaceAround,
                   children: [
@@ -104,12 +99,12 @@ class _SeriesDetailsScreenState extends State<SeriesDetailsScreen> {
                         ),
                       ),
                       child: Text(
-                        AppStrings.aboutSerie,
+                        AppStrings.aboutMovie,
                         style: TextStyle(
                           color: _selectedIndex == 0
                               ? AppColors.primaryText
                               : AppColors.secondaryText,
-                          fontSize: 14,
+                          fontSize: 13,
                         ),
                       ),
                     ),
@@ -170,14 +165,14 @@ class _SeriesDetailsScreenState extends State<SeriesDetailsScreen> {
                   ],
                 ),
                 Padding(
-                  padding: const EdgeInsets.all(10.0),
+                  padding: const EdgeInsets.all(20.0),
                   child: _selectedIndex == 0
-                      ? OverView(series.overview.isEmpty
-                          ? AppStrings.noLanguageSerie
-                          : series.overview)
+                      ? OverView(movie.overview.isEmpty
+                          ? AppStrings.noLanguageMovie
+                          : movie.overview)
                       : _selectedIndex == 1
-                          ? Review(series.voteAverage.toString())
-                          : Cast(series.name),
+                          ? Review(movie.voteAverage.toString())
+                          : Cast(movie.title),
                 ),
               ],
             ),
