+++ lib/presenter/widgets/home/search_bar_app.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:r6_moovie_app/domain/entities/movie.dart';
-import 'package:r6_moovie_app/presenter/pages/movies_details_screen.dart';
+import 'package:r6_moovie_app/presenter/pages/movies/movies_details_screen.dart';
 import 'package:r6_moovie_app/resources/app_colors.dart';
 import 'package:r6_moovie_app/resources/app_strings.dart';
